separate cards to another component
